// contracts/generated/typescript/models/RenderedPageSection.ts
export interface RenderedPageSection {
    /**
     * lead 或稳定顶层 Heading Block ID
     * @type {string}
     * @memberof RenderedPageSection
     */
    key: string;
    /**
     *
     * @type {number}
     * @memberof RenderedPageSection
     */
    position: number;
    /**
     *
     * @type {string}
     * @memberof RenderedPageSection
     */
    headingBlockId?: string;
    /**
     *
     * @type {number}
     * @memberof RenderedPageSection
     */
    level?: number;
    /**
     *
     * @type {string}
     * @memberof RenderedPageSection
     */
    title: string;
    /**
     *
     * @type {number}
     * @memberof RenderedPageSection
     */
    sizeBytes: number;
    /**
     *
     * @type {string}
     * @memberof RenderedPageSection
     */
    revisionId: string;
    /**
     *
     * @type {{ [key: string]: any; }}
     * @memberof RenderedPageSection
     */
    astJson: { [key: string]: any; };
    /**
     *
     * @type {string}
     * @memberof RenderedPageSection
     */
    html: string;
    /**
     *
     * @type {string}
     * @memberof RenderedPageSection
     */
    rendererVersion: string;
}